+++ transformers/ignore_in_tests.go
@@ -1,0 +1,11 @@
+package transformers
+
+import (
+	"reflect"
+)
+
+type IgnoreInTestsTransformer func(field reflect.StructField) bool
+
+func DefaultIgnoreInTestsTransformer(_ reflect.StructField) bool {
+	return false
+}
